+++ src/engine/engine_test.go
@@ -390,6 +390,100 @@ func (self *EngineSuite) TestCountQueryWithGroupByClause(c *C) {
 
 }
 
+// issue #27
+func (self *EngineSuite) TestCountQueryWithGroupByClauseAndNullValues(c *C) {
+	// make the mock coordinator return some data
+	engine := createEngine(c, `
+[
+  {
+    "points": [
+      {
+        "values": [
+          {
+            "string_value": "some_value"
+          },
+				  {
+            "double_value": 1.0
+				  }
+        ],
+        "timestamp": 1381346631000000,
+        "sequence_number": 1
+      },
+      {
+        "values": [
+          {
+            "string_value": "another_value"
+          },
+				  {
+            "double_value": 2.0
+				  }
+        ],
+        "timestamp": 1381346631000000,
+        "sequence_number": 1
+      },
+      {
+        "values": [
+				  null,
+          {
+            "double_value": 3.0
+          }
+        ],
+        "timestamp": 1381346631000000,
+        "sequence_number": 1
+      }
+    ],
+    "name": "foo",
+    "fields": ["column_one", "column_two"]
+  }
+]
+`)
+
+	runQuery(engine, "select count(column_two), column_one from foo group by column_one;", c, `[
+  {
+    "points": [
+      {
+        "values": [
+          {
+            "int64_value": 1
+          },
+          {
+            "string_value": "some_value"
+          }
+        ],
+        "timestamp": 1381346631000000,
+        "sequence_number": 1
+      },
+      {
+        "values": [
+          {
+            "int64_value": 1
+          },
+          {
+            "string_value": "another_value"
+          }
+        ],
+        "timestamp": 1381346631000000,
+        "sequence_number": 1
+      },
+      {
+        "values": [
+          {
+            "int64_value": 1
+          },
+				  null
+        ],
+        "timestamp": 1381346631000000,
+        "sequence_number": 1
+      }
+    ],
+    "name": "foo",
+    "fields": ["count", "column_one"]
+  }
+]
+`)
+
+}
+
 func (self *EngineSuite) BenchmarkFoo(c *C) {
 	counts := map[interface{}]int{}
 	for i := 0; i < c.N; i++ {